changed the circular chart to a column chart. added another column chart to display overall sales of products.

+++ lib/util/prodStats.dart
@@ -1,6 +1,6 @@
 class ProdStats {
   final String name;
-  final int count;
+  final double count;
   final DateTime date;
   ProdStats({required this.name, required this.count, required this.date});
 }
